fix(compilation): Resolve all compilation errors across workspace
Fixed 2 remaining compilation errors:

1. **code_manipulation.rs: Type mismatch i32 vs usize**
   - Changed ExtractFunctionInput.start_line from i32 to usize
   - Changed ExtractFunctionInput.end_line from i32 to usize
   - Now matches extract_function_rust method signature

2. **vfs.rs: Missing create_symlink method**
   - Added create_symlink() method to VirtualFileSystem
   - Creates symbolic link vnodes with proper parent directory handling
   - Follows same pattern as create_directory()

All packages now compile successfully:
- cortex ✓
- cortex-vfs ✓
- cortex-code-analysis ✓
- mcp-sdk ✓
- axon ✓
- dashboard ✓

Workspace compilation: PASS (1m 33s)

+++ cortex/cortex/src/mcp/tools/code_manipulation.rs
@@ -1048,8 +1048,8 @@ impl CodeExtractFunctionTool {
 #[derive(Debug, Deserialize, JsonSchema)]
 struct ExtractFunctionInput {
     source_unit_id: String,
-    start_line: i32,
-    end_line: i32,
+    start_line: usize,
+    end_line: usize,
     function_name: String,
     #[serde(default = "default_before_position")]
     #[allow(dead_code)]
